// citec_app/src/features/utils/interfaces/portfolio-analysis-resp.ts
// Copyright (C) 2021-Present CITEC Inc. <https://citecsolutions.com/>
// All rights reserved
//
// This file is part of CITEC Inc. source code.
// This software framework contains the confidential and proprietary information
// of CITEC Inc., its affiliates, and its licensors. Your use of these
// materials is governed by the terms of the Agreement between your organisation
// and CITEC Inc., and any unauthorised use is forbidden. Except as otherwise
// stated in the Agreement, this software framework is for your internal use
// only and may only be shared outside your organisation with the prior written
// permission of CITEC Inc.
// CITEC Inc. source code can not be copied and/or distributed without the express
// permission of CITEC Inc.
// Generated by https://quicktype.io

export interface PortfolioAnalisysResp {
   fundamentals: Fundamentals
   risk: Risk
   returns: Returns
   capital_gains: CapitalGains
   sectors: Sectors
   factors: Factors
   segment_matrix: SegmentMatrix
   top_holdings: TopHoldings
   concentration: Concentration
   var: Var
   risk_returns: RiskReturns
}

export interface CapitalGains {
   outstanding_balance: number
   unrealized_cap_losses: number
   unrealized_cap_gains: number
   realized_cap_losses: number
   realized_cap_gains: number
   taxes_cost: number
   taxes_rate: number
}

export interface Concentration {
   ticker: string[]
   concentration: number[]
}

export interface Factors {
   factor: string[]
   percentage: number[]
}

export interface Fundamentals {
   total_assets: number
   total_positions: number
   pe: number
   pb: number
   dividend_yield: number
   avg_market_cap: number
}

export interface Returns {
   returns_1M: number
   returns_3M: number
   returns_6M: number
   CAGR: number
   bnch_returns_1M: number
   bnch_returns_3M: number
   bnch_returns_6M: number
   bnch_CAGR: number
}

export interface Risk {
   sharpe: number
   volatility: number
   bnch_sharpe: number
   bnch_volatility: number
   beta: number
   alpha: number
}

export interface RiskReturns {
   ticker: string[]
   returns: number[]
   volatility: number[]
   sharpe: number[]
}

export interface Sectors {
   sector: string[]
   percentage: number[]
   percentage_benchmark: number[]
}

export interface SegmentMatrix {
   style: string[]
   size: string[]
   weight: number[]
}

export interface TopHoldings {
   ticker: string[]
   weight: number[]
   outstanding_balance: number[]
   unrealized_capital_gains: number[]
}

export interface Var {
   quantile: number[]
   value: number[]
}